change the learning_rate to 0.02.
While Adam adapts to the problem, if the orders of magnitude are too different, it may not have time to adjust accordingly. We start overfitting before we can reach a neat solution.
Therefore, for this problem, even 0.02 is a HIGH starting learning rate.
It's a good practice to try 0.001, 0.0001, and 0.00001. If it makes no difference, pick whatever, otherwise it makes sense to fiddle with the learning rate.
Test accuracy : 95.68%